requirements.txt
openai>=1.0.0
openai-agents
litellm
requests>=2.25.0
python-dotenv>=0.19.0 